% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist_app.R
\name{shinyHist}
\alias{shinyHist}
\title{Simple Shiny app to plot a histogram}
\usage{
shinyHist(
  data,
  x,
  title = "Histogram",
  xlab = "x",
  ylab = "y",
  main = "Histogram",
  col = "blue",
  breaks = 10,
  xlim = range(breaks),
  ylim = NULL,
  border = "black",
  lwd = 1,
  lty = 1,
  cex = 1,
  cex.main = 1,
  cex.lab = 1,
  cex.axis = 1,
  cex.sub = 1
)
}
\arguments{
\item{data}{Data frame}

\item{x}{Variable to plot}

\item{title}{Title of the plot}

\item{xlab}{X-axis label}

\item{ylab}{Y-axis label}

\item{main}{Main title}

\item{col}{Color of the bars}

\item{breaks}{Number of breaks}

\item{xlim}{Limits of the x-axis}

\item{ylim}{Limits of the y-axis}

\item{border}{Color of the border}

\item{lwd}{Line width of the border}

\item{lty}{Line type of the border}

\item{cex}{Axis label size}

\item{cex.main}{Main title size}

\item{cex.lab}{Axis label size}

\item{cex.axis}{Axis tick label size}

\item{cex.sub}{Subtitle size}
}
\value{
A Shiny app to plot a histogram.
}
\description{
This is a simple Shiny app to plot a histogram.
Meant as a gentle introduction to Shiny inputs and outputs.
}
\examples{
\dontrun{
data <- data.frame(x = rnorm(100))
shinyHist(data, x = "x")
}

}
\seealso{
\code{\link{hist}}
}
\author{
Jared Andrews
}
